compodb-core/requirements.txt
datafusion==43.1.0
duckdb==1.1.3
filelock
ibis-framework
ibis-substrait
JPype1
protobuf
--extra-index-url https://pypi.fury.io/arrow-nightlies --prefer-binary --pre
pyarrow
pytest
pytest-snapshot
pytest-xdist
substrait
python-dateutil
polars-lts-cpu
fastapi